COM-958 - move export payments
